Split out common test utilities
These may be useful not only in lexer tests, so move them out into
separate modules. Cargo builds tests from 'tests' directory as
standalone binaries so the utilites are put into a common 'utils'
module that needs to be included into each binary. Unfortunately,
this means that utilities will be recompiled for each binary, but
moving them into a crate of their own is too much of a bother.
