+++ src/libsyntax/tests/utils/mod.rs
@@ -1,0 +1,12 @@
+// Copyright (c) 2016, ilammy
+//
+// Licensed under MIT license (see LICENSE file in the root directory).
+// This file may be copied, distributed, and modified only in accordance
+// with the terms specified by this license.
+
+//! Test utilities.
+//!
+//! Here we gather miscellaneous utility code that is used throughout our test harness.
+
+pub mod iter;
+pub mod stubs;
